added spline graph to altitude adjustment dialog
git-svn-id: https://mytourbook.svn.sourceforge.net/svnroot/mytourbook/trunk@970 5a91af82-4d22-0410-8bcb-a41d7c89129f

+++ net.tourbook/src/net/tourbook/data/TourData.java
@@ -592,7 +592,7 @@ public class TourData implements Comparable<Object> {
 	 * contains special points which are displayed in the alti diff graph
 	 */
 	@Transient
-	public int[][]						altiDiffSpecialPoints;
+	public double[][]					altiDiffSpecialPoints;
 
 	/**
 	 * contains the altitude serie which is adjusted at the start
@@ -612,7 +612,6 @@ public class TourData implements Comparable<Object> {
 	 */
 	@Transient
 	private TourData					fMergeSourceTourData;
-
 
 	public TourData() {}
